GCC warning fixes. MinGW compile fixes. Doesn't know a number of VK_... constants.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19586 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/win32/win32kbd.cpp
@@ -403,7 +403,7 @@ bool csWin32KeyboardDriver::HandleKeyMessage (HWND hWnd, UINT message,
 	{
 	  csPrintf ("msg: %s wp: %s lp: %.8x\n", 
 	    Win32KeyMsgNames.StringForIdent (message), GetVKName (wParam), 
-	    lParam);
+	    (uint)lParam);
 	}
 	if (Win32KeyToCSKey (wParam, lParam, raw, cooked, type))
 	{
